+++ app/policies/admin_role_policy.rb
@@ -1,0 +1,2 @@
+class AdminRolePolicy < InstructorPolicy
+end